Refactoring: split MetPostProcessor into its own file and created AurnPostProcessor class
