+++ environmental_data_modules/__init__.py
@@ -1,3 +1,4 @@
 from .environment_module import EnvironmentModule
 from .met_extractor import MetExtractor
-from .post_processor import PostProcessor, MetPostProcessor
+from .post_processor import PostProcessor
+from .met_post_processor import MetPostProcessor
